racing-kings-player/requirements.txt
tensorflow~=2.2.0
matplotlib~=3.2.1
PyQt5~=5.15.0
numpy~=1.18.5
requests~=2.23.0
python-chess~=0.31.2
psutil
filelock
dataclasses
